Update project.py
Added error handling for improper year format. Included handling of years that are not in the database, i.e. requests for a valid year between x and y years. 

Updated handling of sales graph to included adding the report year to the file name, and updated path to save in yearly reports directory.